pyproject.toml
[tool.poetry]
name = "xapps2"
version = "0.1.0"
description = ""
authors = ["Leorio Paradinight <62891774+code-rgb@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
lxml = "^4.6.3"
beautifulsoup4 = "^4.9.3"
pyppeteer = "^0.2.5"
PyYAML = "^5.4.1"
aiohttp = "^3.7.4"
feedparser = "^6.0.8"
uvloop = {version = "^0.16.0", platform = "linux"}
aiodns = "^3.0.0"
Brotli = "^1.0.9"
cchardet = "^2.1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
